+++ common/src/connection/agent.rs
@@ -1,5 +1,4 @@
 use std::{
-    fmt::{Debug, Display},
     pin::Pin,
     task::{Context, Poll},
 };
@@ -8,47 +7,38 @@ use crate::{codec::agent::PpaassAgentConnectionCodec, CommonError, PpaassAgentMe
 
 use futures::{Sink, Stream};
 use pin_project::pin_project;
-use tokio::io::{AsyncRead, AsyncWrite};
+use tokio::net::TcpStream;
 use tokio_util::codec::Framed;
 
 #[non_exhaustive]
 #[pin_project]
-pub struct PpaassAgentConnection<'r, T, R, I>
+pub struct PpaassAgentConnection<R>
 where
-    T: AsyncRead + AsyncWrite + Unpin + Send + Sync + 'static,
-    R: RsaCryptoFetcher + Send + Sync + 'static,
-    I: ToString + Send + Sync + Clone + Display + Debug + 'static,
+    R: RsaCryptoFetcher,
 {
     #[pin]
-    inner: Framed<T, PpaassAgentConnectionCodec<'r, R>>,
-    connection_id: I,
+    inner: Framed<TcpStream, PpaassAgentConnectionCodec<R>>,
+    connection_id: String,
 }
 
-impl<'r, T, R, I> PpaassAgentConnection<'r, T, R, I>
+impl<R> PpaassAgentConnection<R>
 where
-    T: AsyncRead + AsyncWrite + Unpin + Send + Sync + 'static,
     R: RsaCryptoFetcher + Send + Sync + 'static,
-    I: ToString + Send + Sync + Clone + Display + Debug + 'static,
 {
-    pub fn new<'a>(connection_id: I, stream: T, rsa_crypto_fetcher: &'a R, compress: bool, buffer_size: usize) -> PpaassAgentConnection<'r, T, R, I>
-    where
-        'a: 'r,
-    {
+    pub fn new(connection_id: String, stream: TcpStream, rsa_crypto_fetcher: R, compress: bool, buffer_size: usize) -> PpaassAgentConnection<R> {
         let agent_connection_codec = PpaassAgentConnectionCodec::new(compress, rsa_crypto_fetcher);
         let inner = Framed::with_capacity(stream, agent_connection_codec, buffer_size);
         Self { inner, connection_id }
     }
 
-    pub fn get_connection_id(&self) -> &I {
+    pub fn get_connection_id(&self) -> &str {
         &self.connection_id
     }
 }
 
-impl<T, R, I> Sink<PpaassProxyMessage> for PpaassAgentConnection<'_, T, R, I>
+impl<R> Sink<PpaassProxyMessage> for PpaassAgentConnection<R>
 where
-    T: AsyncRead + AsyncWrite + Unpin + Send + Sync + 'static,
     R: RsaCryptoFetcher + Send + Sync + 'static,
-    I: ToString + Send + Sync + Clone + Display + Debug + 'static,
 {
     type Error = CommonError;
 
@@ -73,11 +63,9 @@ where
     }
 }
 
-impl<T, R, I> Stream for PpaassAgentConnection<'_, T, R, I>
+impl<R> Stream for PpaassAgentConnection<R>
 where
-    T: AsyncRead + AsyncWrite + Unpin + Send + Sync + 'static,
     R: RsaCryptoFetcher + Send + Sync + 'static,
-    I: ToString + Send + Sync + Clone + Display + Debug + 'static,
 {
     type Item = Result<PpaassAgentMessage, CommonError>;
