+++ proxy/src/main.rs
@@ -2,8 +2,8 @@ mod common;
 mod config;
 mod crypto;
 mod error;
-mod processor;
 mod server;
+mod transport;
 
 use config::PROXY_CONFIG;
