--- proxy/src/processor/tcp/mod.rs
@@ -1,3 +1,0 @@
-mod destination;
-mod handler;
-pub(in crate::processor) use handler::TcpHandler;
