--- proxy/src/processor/udp/mod.rs
@@ -1,2 +1,0 @@
-mod handler;
-pub(in crate::processor) use handler::UdpHandler;
