+++ proxy/src/transport/tcp/mod.rs
@@ -1,0 +1,3 @@
+mod destination;
+mod handler;
+pub(in crate::transport) use handler::TcpHandler;
