+++ proxy/src/transport/udp/mod.rs
@@ -1,0 +1,2 @@
+mod handler;
+pub(in crate::transport) use handler::UdpHandler;
